feat!: remove the application thumbnail
This is a feature that may have made more sense in the past: a nice
thumbnail to represent the web application we're building for the user.
The idea in the past was to actually get a screenshot of the running
application! https://issues.redhat.com/browse/KONFLUX-1194

.. but, we never brought that into a full feature. The icon is a
randomly chosen one of 9 that is maybe just confusing for the user.

If we could let the user upload and crop an icon, that might be cool: a
custom logo for each application - but, that's a larger effort for
another day: https://issues.redhat.com/browse/KFLUXUI-53

Signed-off-by: Ralph Bean <rbean@redhat.com>