fix(KONFLUX-9575): matrixed Tasks Do Not Appear in Graph View (#400)
* refactor(pipelinerun-graph-utils): enhance matrix logic

Prior to this patch, the matrix logic to detect a matrixed TaskRun was relying
on the 'target-platform' label. This seems to work for the "build-images"
taskRun.
This patch aims to enhance the logic to detect a matrixed TaskRun by checking
if the PipelineTask spec contains the 'matrix' field and the task has multiple
TaskRuns. The displayed name is based off in the taskRun.spec.params data.

* refactor(tekton-results): add records.data.value.spec.params TaskRun field

Update remainingTaskrunFields array to contain the
records.data.value.spec.params field.

* test(pipelinerun-graph-utils): add unit tests

Add unit tests to cover some of the functions implemented in the
pipelinerun-graph-utils file.

* fixup! refactor(pipelinerun-graph-utils): enhance matrix logic

* fixup! test(pipelinerun-graph-utils): add unit tests